requirements.txt
Django==1.7
Markdown==2.5
argparse==1.2.1
django-filter==0.7
djangorestframework==2.4.2
ipython==2.2.0
simplejson==3.6.3
wsgiref==0.1.2
